Initial commit of a fast Social Golfer Problem solver
This doesn't consistently solve the problem, but it consistently gets close in a lot less time.

It's a somewhat general solver allowing you to vary the number of groups, players per group, and the number of rounds to generate.

It also lets you name the golfers, and set global restrictions on pairs of golfers in advance.  The solver reports the conflict scores it computes for each round so you can easily compare results on subsequent runs.

The idea here is not to be mathematically precise, but to provide someone with a practical version of this scheduling problem (for example, a teacher organizing students into different groups) with a fast, almost correct solution that's very easy to use.