Fix player count calculation during csv export
The player count should be: (number of groups) * (group size)
Before this change, it was: (number of rounds) * (group size)

This would only manifest as an issue if the number of rounds was less than the number of groups.
I was trying to export for 14 groups, 8 people, 8 rounds; an error appears on line 174 once the playerIndex exceeds the size of the pre-allocated array (in this case, 64).